add more unit tests
